+++ src/services/api.js
@@ -714,3 +714,19 @@ export function createAttachment({ file }, { onProgress = () => null } = {}) {
     req.send(formData);
   });
 }
+
+export function signOut() {
+  return fetch(`${apiRoot}/v1/session`, postRequestOptions('DELETE'));
+}
+
+export function reissueAuthSession() {
+  return fetch(`${apiRoot}/v1/session/reissue`, postRequestOptions('POST'));
+}
+
+export function listAuthSessions() {
+  return fetch(`${apiRoot}/v1/session/list`, getRequestOptions());
+}
+
+export function closeAuthSessions(ids) {
+  return fetch(`${apiRoot}/v1/session/list`, postRequestOptions('PATCH', { close: ids }));
+}